refactored ws client and ws evt client + cache fix for guild channels and members
