+++ websocket/client.go
@@ -3,217 +3,140 @@ package websocket
 import (
 	"errors"
 	"fmt"
-	"math/rand"
-	"net/http"
-	"runtime"
+	"strconv"
+	"sync"
 	"time"
 
 	"github.com/andersfylling/disgord/constant"
 
-	"github.com/andersfylling/disgord/httd"
+	"github.com/andersfylling/disgord/websocket/cmd"
 	"github.com/andersfylling/disgord/websocket/event"
 	"github.com/andersfylling/disgord/websocket/opcode"
 )
 
 const (
-	// Deprecated
-	maxReconnectTries = 5
+	clientTypeEvent = iota
+	clientTypeVoice
 )
 
-// NewManager creates a new socket client manager for handling behavior and Discord events. Note that this
-// function initiates a go routine.
-func NewClient(config *Config, shardID uint) (client *Client, err error) {
-	if config.TrackedEvents == nil {
-		config.TrackedEvents = &UniqueStringSlice{}
-	}
+// Link is used to establish basic commands to create and destroy a link.
+// See client.Disconnect() and client.Connect() for linking to the Discord servers
+type Link interface {
+	Connect() error
+	Disconnect() error
+}
 
-	var eChan chan<- *Event
-	if config.EventChan != nil {
-		eChan = config.EventChan
-	} else {
-		panic("missing event channel")
-	}
+type connectPermit interface {
+	requestConnectPermit() error
+	releaseConnectPermit() error
+}
 
-	client = &Client{
-		trackedEvents: config.TrackedEvents,
-		eventChan:     eChan,
-	}
-	client.baseClient, err = newBaseClient(config, shardID)
+// newClient ...
+func newClient(config *Config, shardID uint) (c *client, err error) {
+	ws, err := newConn(config.HTTPClient)
 	if err != nil {
 		return nil, err
 	}
-	client.Start()
+
+	c = &client{
+		conf:              config,
+		shutdown:          make(chan interface{}),
+		restart:           make(chan interface{}),
+		ShardID:           shardID,
+		receiveChan:       make(chan *DiscordPacket),
+		emitChan:          make(chan *clientPacket),
+		conn:              ws,
+		ratelimit:         newRatelimiter(),
+		timeoutMultiplier: 1,
+		disconnected:      true,
+		log:               config.Logger,
+		a:                 config.A,
+		behaviors:         map[string]*behavior{},
+		poolDiscordPkt:    config.DiscordPktPool,
+	}
+	c.connectPermit = c
 
 	return
 }
 
-func NewTestClient(config *Config, shardID uint, conn Conn) (*Client, chan interface{}) {
-	s := make(chan interface{})
+// client can be used as a base for other ws clients; voice, event. Note the use of
+// client.ReleasePermit() and client.RequestPermit() in Connect (and then automatically reconnect()).
+// these are used for synchronizing connecting and you must therefore correctly release the permit once you have
+// such that the next shard or yourself can Connect in the future.
+//
+// If you do not care about these. Please overwrite both methods.
+type client struct {
+	sync.RWMutex
+	clientType   int
+	conf         *Config
+	shutdown     chan interface{}
+	restart      chan interface{}
+	lastRestart  int64 //unix
+	restartMutex sync.Mutex
 
-	c := &Client{
-		trackedEvents: config.TrackedEvents,
-		eventChan:     make(chan *Event),
-	}
-	c.Start()
-	go c.receiver()
+	ShardID uint
 
-	return c, s
+	// sending and receiving data
+	ratelimit   ratelimiter
+	receiveChan chan *DiscordPacket
+	emitChan    chan *clientPacket
+	conn        Conn
+
+	// states
+	disconnected      bool
+	haveConnectedOnce bool
+	isReconnecting    bool
+	isReceiving       bool // has the go routine started
+	isEmitting        bool // has the go routine started
+	recEmitMutex      sync.Mutex
+
+	// identify timeout on invalid session
+	// useful in unit tests when you want to drop any actual timeouts
+	timeoutMultiplier int
+
+	log constant.Logger
+
+	// choreographic programming to handle rate limit, reconnects, and connects
+	connectPermit connectPermit
+	K             *K
+	a             A
+
+	// behaviours - optional
+	behaviors map[string]*behavior
+
+	poolDiscordPkt *sync.Pool
 }
 
-// Event is dispatched by the socket layer after parsing and extracting Discord data from a incoming packet.
-// This is the data structure used by Disgord for triggering handlers and channels with an event.
-type Event struct {
-	Name string
-	Data []byte
+type behaviorActions map[interface{}]actionFunc
+type actionFunc func(interface{}) error
+type behavior struct {
+	addresses string
+	actions   behaviorActions
 }
 
-// Config ws
-// TODO: remove shardID, such that this struct can be reused for every shard
-type Config struct {
-	// BotToken Discord bot token
-	BotToken string
-
-	// HTTPClient custom http client to support the use of proxy
-	HTTPClient *http.Client
-
-	// ChannelBuffer is used to set the event channel buffer
-	ChannelBuffer uint
-
-	// Endpoint for establishing socket connection. Either endpoints, `Gateway` or `Gateway Bot`, is used to retrieve
-	// a valid socket endpoint from Discord
-	Endpoint string
-
-	// Encoding make sure we support the correct encoding
-	Encoding string
-
-	// Version make sure we support the correct Discord version
-	Version int
-
-	// TrackedEvents holds a list of predetermined events that should not be ignored.
-	// This is especially useful for creating multiple shards, to reuse the same slice
-	TrackedEvents *UniqueStringSlice
-
-	// EventChan can be used to inject a channel instead of letting the ws client construct one
-	// useful in sharding to avoid complicated patterns to handle N channels.
-	EventChan chan<- *Event
-
-	A A
-
-	Logger constant.Logger
-
-	// for identify packets
-	Browser             string
-	Device              string
-	GuildLargeThreshold uint
-	ShardCount          uint
+func (c *client) addBehavior(b *behavior) {
+	c.behaviors[b.addresses] = b
 }
 
-type Client struct {
-	*baseClient
-	ReadyCounter uint
+const (
+	discordOperations string = "discord-ops"
+	heartbeating      string = "heartbeats"
+)
 
-	eventChan        chan<- *Event
-	trackedEvents    *UniqueStringSlice
-	heartbeatLatency time.Duration
-
-	sessionID string
-	trace     []string
-}
-
-var _ constant.Logger = (*Client)(nil)
-
-// Connect establishes a socket connection with the Discord API
-func (c *Client) Connect() (err error) {
-	return c.connect()
-}
-
-// HeartbeatLatency get the time diff between sending a heartbeat and Discord replying with a heartbeat ack
-func (c *Client) HeartbeatLatency() (duration time.Duration, err error) {
-	duration = c.heartbeatLatency
-	if duration == 0 {
-		err = errors.New("latency not determined yet")
-	}
-
-	return
-}
-
-// RegisterEvent tells the socket layer which event types are of interest. Any event that are not registered
-// will be discarded once the socket info is extracted from the event.
-func (c *Client) RegisterEvent(event string) {
-	c.trackedEvents.Add(event)
-}
-
-// RemoveEvent removes an event type from the registry. This will cause the event type to be discarded
-// by the socket layer.
-func (c *Client) RemoveEvent(event string) {
-	c.trackedEvents.Remove(event)
-}
-
-func (c *Client) Start() {
-	go c.operationHandlers()
-}
-
-func (c *Client) Shutdown() (err error) {
-	close(c.shutdown)
-	c.Disconnect()
-	return
-}
-
-func (c *Client) eventHandler(p *discordPacket) {
-	// discord events
-	// events that directly correlates to the socket layer, will be dealt with here. But still dispatched.
-
-	// increment the sequence number for each event to make sure everything is synced with discord
-	c.Lock()
-	c.sequenceNumber++
-
-	// validate the sequence numbers
-	if p.SequenceNumber != c.sequenceNumber {
-		c.Info(fmt.Sprintf("websocket sequence numbers missmatch, forcing reconnect. Got %d, wants %d", p.SequenceNumber, c.sequenceNumber))
-		c.sequenceNumber--
-		c.Unlock()
-		go c.reconnect()
-		return
-	}
-	c.Unlock()
-
-	if p.EventName == event.Ready {
-
-		// always store the session id & update the trace content
-		ready := readyPacket{}
-		err := httd.Unmarshal(p.Data, &ready)
-		if err != nil {
-			c.Error(err.Error())
+func (c *client) start() {
+	for k := range c.behaviors {
+		switch k {
+		case discordOperations:
+			go c.operationHandlers()
 		}
-
-		c.Lock()
-		c.sessionID = ready.SessionID
-		c.trace = ready.Trace
-		c.ReadyCounter++
-		c.Unlock()
-	} else if p.EventName == event.Resume {
-		// eh? debugging.
-		// TODO
-	} else if p.Op == opcode.DiscordEvent && !c.eventOfInterest(p.EventName) {
-		return
 	}
-
-	// dispatch event
-	c.eventChan <- &Event{
-		Name: p.EventName,
-		Data: p.Data,
-	}
-} // end eventHandler()
-
-func (c *Client) eventOfInterest(name string) bool {
-	return c.trackedEvents.Exists(name)
 }
 
-// operation handler demultiplexer
-func (c *Client) operationHandlers() {
+// operation handler de-multiplexer
+func (c *client) operationHandlers() {
 	c.Debug("Ready to receive operation codes...")
 	for {
-		var p *discordPacket
+		var p *DiscordPacket
 		var open bool
 		select {
 		case p, open = <-c.Receive():
@@ -227,132 +150,442 @@ func (c *Client) operationHandlers() {
 			return
 		}
 
-		// new packet that must be handled by it's Discord operation code
-		switch p.Op {
-		case opcode.DiscordEvent:
-			c.eventHandler(p)
-		case opcode.Reconnect:
-			c.Info("Discord requested a reconnect")
-			go c.reconnect()
-		case opcode.InvalidSession:
-			// invalid session. Must respond with a identify packet
-			c.Info("Discord invalidated session")
-			go func() {
-				// session is invalidated, reset the sequence number
-				c.sequenceNumber = 0
-
-				rand.Seed(time.Now().UnixNano())
-				delay := rand.Intn(4) + 1
-				delay *= c.timeoutMultiplier
-				randomDelay := time.Second * time.Duration(delay)
-
-				// This ignores the identify rate limit of 1/5s, because of the documentation stating:
-				//  It's also possible that your client cannot reconnect in time to resume, in which case
-				//  the client will receive a Opcode 9 Invalid Session and is expected to wait a random
-				//  amount of time—between 1 and 5 seconds—then send a fresh Opcode 2 Identify.
-				<-time.After(randomDelay)
-				err := sendIdentityPacket(c)
-				if err != nil {
-					c.Error(err.Error())
-				}
-			}()
-		case opcode.Heartbeat:
-			// https://discordapp.com/developers/docs/topics/gateway#heartbeating
-			_ = c.Emit(event.Heartbeat, c.sequenceNumber)
-		case opcode.Hello:
-			// hello
-			helloPk := &helloPacket{}
-			err := httd.Unmarshal(p.Data, helloPk)
+		if action, defined := c.behaviors[discordOperations].actions[p.Op]; defined {
+			err := action(p)
 			if err != nil {
-				c.Debug(err.Error())
+				c.Error(err.Error())
 			}
-			c.Lock()
-			c.baseClient.heartbeatInterval = helloPk.HeartbeatInterval
-			c.Unlock()
-
-			c.sendHelloPacket()
-		case opcode.HeartbeatAck:
-			// heartbeat received
-			c.Lock()
-			c.lastHeartbeatAck = time.Now()
-			c.Unlock()
-		default:
-			// unknown
-			c.Debug(fmt.Sprintf("Unknown operation: %+v\n", p))
 		}
+
+		// see receiver() for creation/Get()
+		c.poolDiscordPkt.Put(p)
 	}
 }
 
-func (c *Client) sendHelloPacket() {
-	// TODO, this might create several idle goroutines..
-	go c.pulsate()
+//////////////////////////////////////////////////////
+//
+// SHARD synchronization & rate limiting
+//
+//////////////////////////////////////////////////////
 
-	// if this is a new connection we can drop the resume packet
-	if c.sessionID == "" && c.sequenceNumber == 0 {
-		err := sendIdentityPacket(c)
-		if err != nil {
-			c.Error(err.Error())
+func (c *client) requestConnectPermit() error {
+	c.Debug("trying to get Connect permission")
+	b := make(B)
+	defer close(b)
+	c.a <- b
+	c.Debug("waiting")
+	var ok bool
+	select {
+	case c.K, ok = <-b:
+		if !ok || c.K == nil {
+			c.Debug("unable to get Connect permission")
+			return errors.New("channel closed or K was nil")
+		}
+		c.Debug("got Connect permission")
+	case <-c.shutdown:
+	}
+
+	return nil
+}
+
+func (c *client) releaseConnectPermit() error {
+	if c.K == nil {
+		return errors.New("K has not been granted yet")
+	}
+
+	c.K.Release <- c.K
+	c.K = nil
+	return nil
+}
+
+//////////////////////////////////////////////////////
+//
+// LOGGING
+//
+//////////////////////////////////////////////////////
+
+func (c *client) getLogPrefix() string {
+	if c.clientType == clientTypeVoice {
+		return "[ws, voice] "
+	}
+
+	// [ws, event, shard:0]
+	return "" +
+		"[ws, " +
+		"event, " +
+		"shard:" +
+		strconv.FormatUint(uint64(c.ShardID), 10) +
+		"] "
+}
+
+func (c *client) Info(msg string) {
+	if c.log != nil {
+		c.log.Info(c.getLogPrefix() + msg)
+	}
+}
+func (c *client) Debug(msg string) {
+	if c.log != nil {
+		c.log.Debug(c.getLogPrefix() + msg)
+	}
+}
+func (c *client) Error(msg string) {
+	if c.log != nil {
+		c.log.Error(c.getLogPrefix() + msg)
+	}
+}
+
+var _ constant.Logger = (*client)(nil)
+
+//////////////////////////////////////////////////////
+//
+// LINKING: CONNECTING / DISCONNECTING / RECONNECTING
+//
+//////////////////////////////////////////////////////
+
+// Connect establishes a socket connection with the Discord API
+func (c *client) Connect() (err error) {
+	c.Lock()
+	defer c.Unlock()
+
+	// c.conn.Disconnected can always tell us if we are disconnected, but it cannot with
+	// certainty say if we are connected
+	if !c.disconnected {
+		err = errors.New("cannot Connect while a connection already exist")
+		return
+	}
+
+	if c.conf.Endpoint == "" {
+		panic("missing websocket endpoint. Must be set before constructing the sockets")
+		//c.conf.Endpoint, err = getGatewayRoute(c.conf.HTTPClient, c.conf.Version)
+		//if err != nil {
+		//	return
+		//}
+	}
+
+	err = c.connectPermit.requestConnectPermit()
+	if err != nil {
+		err = errors.New("unable to get permission to Connect. Err: " + err.Error())
+		return
+	}
+
+	// establish ws connection
+	err = c.conn.Open(c.conf.Endpoint, nil)
+	if err != nil {
+		if !c.conn.Disconnected() {
+			c.conn.Close()
+			// TODO: logging
 		}
 		return
 	}
 
-	c.RLock()
-	token := c.conf.BotToken
-	session := c.sessionID
-	sequence := c.sequenceNumber
-	c.RUnlock()
+	// we can now interact with Discord
+	c.haveConnectedOnce = true
+	c.disconnected = false
+	go c.receiver()
+	go c.emitter()
 
-	err := c.Emit(event.Resume, struct {
-		Token      string `json:"token"`
-		SessionID  string `json:"session_id"`
-		SequenceNr uint   `json:"seq"`
-	}{token, session, sequence})
-	if err != nil {
-		c.Error(err.Error())
-	}
-
-	err = c.connectPermit.releaseConnectPermit()
-	if err != nil {
-		err = errors.New("unable to release connection permission. Err: " + err.Error())
-		c.Error(err.Error())
-	}
+	return
 }
 
-func sendIdentityPacket(c *Client) (err error) {
-	// https://discordapp.com/developers/docs/topics/gateway#identify
-	identityPayload := struct {
-		Token          string      `json:"token"`
-		Properties     interface{} `json:"properties"`
-		Compress       bool        `json:"compress"`
-		LargeThreshold uint        `json:"large_threshold"`
-		Shard          *[2]uint    `json:"shard,omitempty"`
-		Presence       interface{} `json:"presence,omitempty"`
-	}{
-		Token: c.conf.BotToken,
-		Properties: struct {
-			OS      string `json:"$os"`
-			Browser string `json:"$browser"`
-			Device  string `json:"$device"`
-		}{runtime.GOOS, c.conf.Browser, c.conf.Device},
-		LargeThreshold: c.conf.GuildLargeThreshold,
-		Shard:          &[2]uint{c.ShardID, c.conf.ShardCount},
-		// Presence: struct {
-		// 	Since  *uint       `json:"since"`
-		// 	Game   interface{} `json:"game"`
-		// 	Status string      `json:"status"`
-		// 	AFK    bool        `json:"afk"`
-		// }{Status: "online"},
+// Disconnect disconnects the socket connection
+func (c *client) Disconnect() (err error) {
+	c.Lock()
+	defer c.Unlock()
+	if c.conn.Disconnected() || !c.haveConnectedOnce {
+		c.disconnected = true
+		err = errors.New("already disconnected")
+		return
 	}
 
-	err = c.Emit(event.Identify, &identityPayload)
+	// use the emitter to dispatch the close message
+	err = c.conn.Close()
 	if err != nil {
-		return err
+		c.Error(err.Error())
+	}
+	// c.Emit(event.Close, nil)
+	// dont use emit, such that we can call shutdown at the same time as Disconnect (See Shutdown())
+	c.disconnected = true
+
+	// close connection
+	<-time.After(time.Second * 1 * time.Duration(c.timeoutMultiplier))
+	return
+}
+
+func (c *client) lockReconnect() bool {
+	c.restartMutex.Lock()
+	defer c.restartMutex.Unlock()
+
+	now := time.Now().UnixNano()
+	locked := (now - c.lastRestart) < (time.Second.Nanoseconds() / 2)
+
+	if !locked && !c.isReconnecting {
+		c.lastRestart = now
+		c.isReconnecting = true
+		return true
 	}
 
-	err = c.connectPermit.releaseConnectPermit()
-	if err != nil {
-		err = errors.New("unable to release connection permission. Err: " + err.Error())
+	return false
+}
+
+func (c *client) unlockReconnect() {
+	c.restartMutex.Lock()
+	defer c.restartMutex.Unlock()
+
+	c.isReconnecting = false
+}
+
+func (c *client) reconnect() (err error) {
+	// make sure there aren't multiple reconnect processes running
+	if !c.lockReconnect() {
+		return
+	}
+	defer c.unlockReconnect()
+
+	c.Debug("is reconnecting")
+
+	c.restart <- 1
+	_ = c.Disconnect()
+
+	var try uint
+	var delay time.Duration = 3 // seconds
+	for {
+		c.Debug(fmt.Sprintf("EventReconnect attempt #%d\n", try))
+		err = c.Connect()
+		if err == nil {
+			c.Info("successfully reconnected")
+			break
+		}
+
+		err = c.connectPermit.releaseConnectPermit()
+		if err != nil {
+			err = errors.New("unable to release connection permission. Err: " + err.Error())
+			c.Info(err.Error())
+		}
+
+		c.Info(fmt.Sprintf("reconnect failed, trying again in N seconds; N =  %d", uint(delay)))
+		c.Info(err.Error())
+
+		// wait N seconds
+		select {
+		case <-time.After(delay * time.Second):
+			delay += 4 + time.Duration(try*2)
+		case <-c.shutdown:
+			return
+		}
+
+		if uint(delay) > 5*60 {
+			delay = 60
+		}
 	}
 
 	return
 }
+
+var _ Link = (*client)(nil)
+
+//////////////////////////////////////////////////////
+//
+// EMITTING / DISPATCHING
+//
+//////////////////////////////////////////////////////
+
+// Emit is used by DisGord users for dispatching a socket command to the Discord Gateway.
+func (c *client) Emit(command string, data interface{}) (err error) {
+	if !c.haveConnectedOnce {
+		return errors.New("race condition detected: you must Connect to the socket API/Gateway before you can send gateway commands")
+	}
+
+	op := ^uint(0)
+	// TODO: refactor command and event name to avoid conversion (?)
+	if c.clientType == clientTypeVoice {
+		switch command {
+		case cmd.VoiceSpeaking:
+			op = opcode.VoiceSpeaking
+		case cmd.VoiceIdentify:
+			op = opcode.VoiceIdentify
+		case cmd.VoiceSelectProtocol:
+			op = opcode.VoiceSelectProtocol
+		case cmd.VoiceHeartbeat:
+			op = opcode.VoiceHeartbeat
+		case cmd.VoiceResume:
+			op = opcode.VoiceResume
+		}
+	} else if c.clientType == clientTypeEvent {
+		switch command {
+		case event.Shutdown:
+			op = opcode.Shutdown
+		case event.Close:
+			op = opcode.Close
+		case event.Heartbeat:
+			op = opcode.EventHeartbeat
+		case event.Identify:
+			op = opcode.EventIdentify
+		case event.Resume:
+			op = opcode.EventResume
+		case cmd.RequestGuildMembers:
+			op = opcode.EventRequestGuildMembers
+		case cmd.UpdateVoiceState:
+			op = opcode.EventVoiceStateUpdate
+		case cmd.UpdateStatus:
+			op = opcode.EventStatusUpdate
+		}
+	}
+	if op == ^uint(0) {
+		err = errors.New("unsupported command: " + command)
+		return
+	}
+
+	accepted := c.ratelimit.Request(command)
+	if !accepted {
+		return errors.New("rate limited")
+	}
+
+	c.emitChan <- &clientPacket{
+		Op:   op,
+		Data: data,
+	}
+	return
+}
+
+func (c *client) lockEmitter() bool {
+	c.recEmitMutex.Lock()
+	defer c.recEmitMutex.Unlock()
+
+	if !c.isEmitting {
+		c.isEmitting = true
+		return true
+	}
+
+	return false
+}
+
+func (c *client) unlockEmitter() {
+	c.recEmitMutex.Lock()
+	defer c.recEmitMutex.Unlock()
+
+	c.isEmitting = false
+}
+
+// emitter holds the actually dispatching logic for sending data to the Discord Gateway.
+// client#Emit depends on this.
+func (c *client) emitter() {
+	if !c.lockEmitter() {
+		c.Error("tried to start another websocket emitter go routine")
+		return
+	}
+	defer c.unlockEmitter()
+
+	for {
+		var msg *clientPacket
+		var open bool
+
+		select {
+		case <-c.shutdown:
+			// c.connection got closed
+		case msg, open = <-c.emitChan:
+		}
+		if !open || (msg.Data == nil && (msg.Op == opcode.Shutdown || msg.Op == opcode.Close)) {
+			// TODO: what if we get a connection error, how do we restart?
+			err := c.Disconnect()
+			if err != nil {
+				c.Error(err.Error())
+			}
+			c.Debug("closing emitter")
+			return
+		}
+		var err error
+
+		// save to file
+		// build tag: disgord_diagnosews
+		saveOutgoingPacket(c, msg)
+
+		err = c.conn.WriteJSON(msg)
+		if err != nil {
+			c.Error(err.Error())
+		}
+	}
+}
+
+//////////////////////////////////////////////////////
+//
+// RECEIVING
+//
+//////////////////////////////////////////////////////
+
+// Receive returns the channel for receiving Discord packets
+func (c *client) Receive() <-chan *DiscordPacket {
+	return c.receiveChan
+}
+
+func (c *client) lockReceiver() bool {
+	c.recEmitMutex.Lock()
+	defer c.recEmitMutex.Unlock()
+
+	if !c.isReceiving {
+		c.isReceiving = true
+		return true
+	}
+
+	return false
+}
+
+func (c *client) unlockReceiver() {
+	c.recEmitMutex.Lock()
+	defer c.recEmitMutex.Unlock()
+
+	c.isReceiving = false
+}
+
+func (c *client) receiver() {
+	if !c.lockReceiver() {
+		c.Error("tried to start another websocket receiver go routine")
+		return
+	}
+	defer c.unlockReceiver()
+
+	for {
+		packet, err := c.conn.Read()
+		if err != nil {
+			c.Debug("closing receiver")
+			return
+		}
+
+		// parse to gateway payload object
+		// see operationHandler for return/Put()
+		evt := c.poolDiscordPkt.Get().(*DiscordPacket)
+		evt.reset()
+		err = evt.UnmarshalJSON(packet)
+		if err != nil {
+			c.Error(err.Error())
+			continue
+		}
+
+		// save to file
+		// build tag: disgord_diagnosews
+		saveIncomingPacker(c, evt, packet)
+
+		// notify listeners
+		c.receiveChan <- evt
+
+		// check if application has closed
+		select {
+		case <-c.shutdown:
+			return
+		default:
+		}
+	}
+}
+
+func (c *client) Shutdown() (err error) {
+	close(c.shutdown)
+	c.Disconnect()
+	return
+}
+
+//////////////////////////////////////////////////////
+//
+// HEARTBEAT / PULSATING
+//
+//////////////////////////////////////////////////////
+
+// TODO - is it worth it?
